% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srilankalatex.R
\docType{data}
\name{srilankalatex}
\alias{srilankalatex}
\title{Sri Lanka Latex Dataset}
\format{
A data frame with 15 variables:
\describe{
\item{Year}{Year of observation.}
\item{Month}{Month of observation.}
\item{Zone}{Agro-climatic zone.}
\item{RPC}{Regional Plantation Company.}
\item{Estate}{Estate name.}
\item{Division}{Division name.}
\item{Latex}{Latex production.}
\item{Rainfall}{Monthly rainfall.}
\item{Tappable_trees}{Number of tappable trees.}
\item{Immature_trees}{Number of immature trees.}
\item{Uprooted_trees}{Number of uprooted trees.}
\item{Fertilizer_applied_trees}{Trees receiving fertilizer.}
\item{Stimulant_applied_trees}{Trees receiving stimulant.}
\item{Pests_and_diseases_infected_trees}{Trees infected by pests and diseases.}
\item{Labour_turnover}{Labour turnover.}
}
}
\usage{
srilankalatex
}
\description{
Monthly latex production and plantation management data from Sri Lanka.
}
\keyword{datasets}
